requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attr==0.3.1
attrs==25.4.0
audioread==3.1.0
azure-core==1.37.0
azure-storage-blob==12.27.1
black==26.1a1
bleach==5.0.1
blinker==1.9.0
boto3==1.42.17
botocore==1.42.17
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
datamodel-code-generator==0.26.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.15
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==3.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-spectacular==0.28.0
email-validator==2.3.0
executing==2.2.1
expiringdict==1.2.2
Faker==39.0.0
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gast==0.7.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.5.0
google-cloud-logging==3.13.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.1
inflect==5.6.2
inflection==0.5.1
ipykernel==7.0.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsf==0.11.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.13.0
kiwisolver==1.4.9
label-studio==1.22.0
label-studio-sdk==2.0.16
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.46.0
lockfile==0.12.2
lxml==6.0.2
lxml_html_clean==0.4.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.2.6
openai==1.109.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
opt_einsum==3.4.0
optree==0.18.0
ordered-set==4.0.2
packaging==25.0
pandas==2.2.3
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
polars==1.36.1
polars-runtime-32==1.36.1
pooch==1.8.2
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytokens==0.3.0
pytz==2022.7.1
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
redis==5.2.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-mock==1.12.1
rich==14.2.0
rpds-py==0.30.0
rq==2.6.1
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
semver==3.0.4
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.5
stack-data==0.6.3
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tldextract==5.3.1
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
ultralytics==8.3.247
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.2
user-agents==2.2.0
uuid_utils==0.12.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.4
wget==3.2
word2number==1.1
wrapt==2.0.1
xmljson==0.2.1
zipp==3.23.0
